% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vpa_result}
\alias{vpa_result}
\title{VPA result calculated using \code{frasyr::vpa()}}
\format{WRITE DATA FORMAT HERE}
\usage{
data(vpa_result)
}
\description{
VPA result calculated using \code{frasyr::vpa()}
}
\keyword{datasets}
